refactor: group dark mode logic to useDark
